Software AX mixer on SDL3, fighter attrs as disc data, sign-extension sweep, HUD/collision/ARQ fixes
- src/pc/audio.c: DSP-ADPCM/PCM voice mixer at 32kHz/5ms frames driving the synth callback;
  MELEE_AUDIO_DUMP for verification
- aurora: ARQPostRequest fills the request, aurora_aram_base(), CARD log noise -> debug
- 8 fighters' DatAttrs annotated DISC_STRUCT; ftCollisionBoxDisc helpers
- ~200 int->pointer casts made zero-extending or retyped (sign-extension of 0x80xxxxxx)
- ifstatus HUD: JObj child walk, anim-chain depth; ifstock/iftime DiscU32 tables
- mplib struct pair overlay, lbarq typed owner, volatile spin reads, player.c overlay removed
- CardState layout-exact; sislib record size; GameCube-offset hacks in gr/mn/ty replaced

+++ src/melee/gm/gmevent.c
@@ -609,7 +609,7 @@ void onEnterVs(GameModeState* arg0)
             gm_8016A434();
         }
         if (event_info[level]->x4 != NULL) {
-            gm_8016A404((s32) event_info[level]->x4);
+            gm_8016A404(event_info[level]->x4);
         }
         if (EV_BONUS(LV(levels, level))->x15 != 0) {
             gm_8016A424(EV_BONUS(LV(levels, level))->x15);
